Tighten up the safety of StarsectorAPIInteractor
Turns out, all these Global.getSector().campaignUI fields can be null.
So, tighten up the nullsafety around them, extract that into a new method that
returns an Optional, and use the method instead of directly asking for them in a non-nullable way.

Now, there is a decent system which can be neither Refit nor ExoticaTech screen, and in that case
the HullmodExoticHandler will scream with errors. Lets cross that bridge if/when we get to it.

Ditto for nullsafety on the Global.getSector().campaignUI.currentCoreTab == CoreUITabId.REFIT part.