Extend the "warning" idea into actually also preventing removal from Refit screen
Well, the Modification got a new method blockRemovalFromRefitScreen() which defaults to false.

HullmodExotic is overriding it, and setting it to true.

ExoticMethodsUIPlugin is doing most of the work, where after showing the "don't use" / "installed"
it's also showing the method buttons, so a small hack has been put in there to disable the recovery
methods if we're running from refit screen (and the exotic should warn and block removal from refit)