Latest JAR - RC9
This version fixes the NPE problem that BaneOfOrion reported for 0.98 among fixing other small tidbits here and there.

There are a few outstanding problems which were identified:

1) blocking (or warning users, but better to block) to remove from refit

However, for that I'm thinking of putting in a new string/tooltip somewhere below the methods,
which will show a "preapply string" which says what the method does on-hover, which should be
painted red and say "do not use from refit" for hullmod exotics in fashion similar to whats tried before.

2) fixing HullmodExoticHandler a bit

currently, if effects are shared to children, the "remove from all children" also removes it from the parent
fleet member, which should not be the case, so the CheckAndRemoveFromMemberModule flow is never executed
because it's already been removed from there. Obviously, the parentFleetMember should be skipped during the
CheckAndRemoveFromAllChildModulesVariants (as well as CheckAndInstallOnAllChildModulesVariants for good measure)

3) really seeing why removing the AlphaSubcore from exotica screen isn't always removing it completely from all modules.

4) well, this is a non-issue, the GuardianShield's flux debuff penalty won't be moved to onBeforeShipCreation() since
it's only active while the GuardianShield is active, and is removed while the shield is inactive.