Parametrize removing from installed/expected lists in STRICT mode
Aka put in more flags in the HullmodExoticHandler.

I remember something being problematic with ships that have two different modules using
the same hull (or variant) ID, and that causing a problem in the lookup map when the first module
gets removed and the bonus remaining on the second module which was, unfortunatelly, already removed
from the map.

So lets parametrize that so we can play around with that without changing the code too much.

However, I'm also noticing another problematic thing - regarding the following log spam:
```
4818142 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnAllChildModulesVariants() from neither Refit or ExoticaTech screens! Bailing out...
4818142 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnMemberModule() from neither Refit or ExoticaTech screens! Bailing out...
4818143 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnAllChildModulesVariants() from neither Refit or ExoticaTech screens! Bailing out...
4818144 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnMemberModule() from neither Refit or ExoticaTech screens! Bailing out...
4818144 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnAllChildModulesVariants() from neither Refit or ExoticaTech screens! Bailing out...
4818144 [Thread-2] ERROR exoticatechnologies.hullmods.exotics.HullmodExoticHandler  - Illegal state detected, tried to CheckAndInstallOnMemberModule() from neither Refit or ExoticaTech screens! Bailing out...
```

Namely, entering or leaving a planet tries installing modules on all of the ships *again*, and this hits a problem with the HullmodExotics
which expect that to happen just from the refit screen or the exotica screen. Perhaps parametrizing this as well will prove useful before
I try playing with those switches a bit.

Right now, installing the AlphaSubcore on the main module and removing it later on (either from Exotica or Refit screens) just does not work.