Set the ALLOW_INTERACTION_OUTSIDE_EXOTICA_OR_REFIT flag to true
Well, there's far less spam in the log, but overall - it's still completely broken in the edge-case when
Alpha Subcore isn't the last exotic to be removed from a child module.

Installing works fine, uninstalling won't work well for child modules where e.g. Alpha Subcore isn't the
last exotic to be removed from it (in case there will be exotics leftover after if was supposed to be removed).

However, the situation can be completely remedied by the user by following a certain set of steps,
albeit at a price of either Story Points or the Exotic chip itself (when destroying it)

The better wall of text follows.

**NOTE:**

The install/uninstall for HullmodExotics and HullmodExotics doesn't play very nice with other exotics. (**Alpha Subcore** is one of them, **Daemon Core** is the other one). To be more specific, the issue seems to be very related to trying to (auto)remove them from modules that still have exotics on them.

To completely avoid the issue, try to make sure that Alpha Subcore is the last Exotic to be removed on all modules; which means - remove all exotics from the "offending" module(s) before removing the HullmodExotic from either the Refit or Exotica screen

*WHEN THERE WERE EXOTICS (MAYBE EVEN UPGRADES) ON MODULES BEFORE INSTALLING THEM:*

Looking at a 3-module ship, in the following example:
- installing Nanotech Armor (irrelevant) in first child module
- installing Equalizer Core/Spooled Feeders (irrelevant) in the second child module
- installing Alpha Subcore on the main module

will properly reduce OP costs to all modules on the ship (even from refit) or rather, it will add itself to all modules, given that there is room for it on the child modules (won't exceed maximum number of exoticas)

But playing the game / leaving the planet and coming back to it
- uninstalling Alpha Subcore from the main module (via Refit screen***** or Exotica screen)

**should** properly reduce OP costs across all modules. **HOWEVER** chances are that it won't work properly every time, in which case the way to "fix" ships is to, unfortunatelly, ensure the Alpha Subcore is the *last* modification (exotic/upgrade) removed from the module.

***** - as advised, try not doing it from refit. However, while I was typing it out, I wanted to document the defect exactly, and it worked perfectly.

_at the time of documenting, the expected "failure scenario" did not fail at all and actually worked perfectly; both the installation and uninstallation have been done through Refit screen_

**UPON RETRYING:** (from Refit screen)
The Alpha Subcore remained active (not uninstalled) from the  child modules.

In that case, this set of steps (relevant to the same example) *should* fix the child modules:
- uninstall Nanotech Armor (irrelevant) from the first child module using Refit screen
- uninstall Equalizer Core / Spooled Feeders (irrelevant) from the second child module using Refit screen
- change tab to cargo or something else
- uninstall Alpha Subcore from the main module using the Exotica screen

*WHEN THERE WEREN'T EXOTICS/UPGRADES ON MODULES BEFORE INSTALLING THEM:*

- install Alpha Subcore on the main module using Exotica screen
- install Spooled Feeders and Nanotech Armor in the first child module

When Alpha Subcore is installed first,  it will surely properly reduce OP on all modules on the ship (Refit is flaky here, but tends to work. Still, always prefer Exotica screen for these two exotics), followed by installing exotics on other modules. However, after
- Uninstalling Alpha Subcore from the main module (via Refit screen or Exotica screen)
- checking the ship in refit

Chances aren't small for the Alpha Subcore to remain on the module where Spooled Feeders were installed. However, in that case, the remedy steps are **identical** as to the other situation (read above):
- remove all exoticas from the child module
- install Alpha Subcore in the main module
- uninstall Alpha Subcore from the main module
**should** completely remove the "bonus OP" bonus and the OP on the modules should be restored.