Latest JAR - RC10 and recap
Looking at "what still needs to be done" from one of the previous commit messages:

>    There are a few outstanding problems which were identified:
>
>    1) blocking (or warning users, but better to block) to remove from refit

Decided to actually *not* block, because Refit works flaky (and not downright not-working)
but actually works fine most of the time

>    2) fixing HullmodExoticHandler a bit

Was kinda done in the last branch, wasn't it? If nothing, the defect was clearly identified and
a remedy was found. The log spams much less now than before, now that I've allowed interactions
outside of Refit or Exotica screen.

>    3) really seeing why removing the AlphaSubcore from exotica screen isn't always removing it completely from all modules.

I believe the root cause of this was what the latest branch #53 tried fixing. There were surely leftover exotics in there,
or some such nonsense, but I don't feel like looking into this or blocking further development over this tiny thing anymore.

Enough is enough. It's a known issue, it's expected to happen, a remedy was found, that's that.

>    4) well, this is a non-issue, the GuardianShield's flux debuff penalty won't be moved to onBeforeShipCreation() since
>    it's only active while the GuardianShield is active, and is removed while the shield is inactive.

There was nothing to do about this either.

The only **real** thing left to do is stress-test this a bit in my own game, and see if there are multiple guardian shields spawning
and then tackling that again (without redoing the whole framework again, since that's already done now)