Touch-up FullMetalSalvo
While there was nothing inherently wrong with it, I did get confused about the whole
"works on whole ship", expecting it to light up all modules at the same time.

Turns out, each module activates on it's own when appropriate, thus we don't need
to apply the AfterImage to the whole ship when that happens, just let each module light
up when it activates and that's that.

Additionally, the "bonus" and "malus" parts were extracted into separate methods, as well as
making the "projectile speed" and "damage boost" bonuses take positive mult into their
actual calculation instead of just using a constant for it. Same thing for "rate of fire" malus,
which takes in the negative multiplier.

That being said, the "positives" are:
- projectile speed boost
- damage boost
- duration

while cooldown and rate of fire maluses are "negatives". The coefficient is neither, and will remain a flat
33% until further notice/revision because when something is applied to it, it quickly gets out of hand with
e.g. Pure type, which will then apply the positive mult twice, resulting in a 6.25x increase (e.g. 206% instead of 33%)

Visual flair on "improved bullets" (bullets coming from Metal Salvo-activated modules) was considered but might
happen at a later point as well.