Add NaN-guards to PureType - part 2
The previous implementation just returned 0f in case of NaN, now we do something
a bit smarter, where we either return a 0f if numerator was bigger than 0 and the denominator was 0.

in case where both numerator and denominator was 0, while that should have been 0,
it probably resulted in a NaN, so we return 1f because they were equal.

UPDATE: also added logging in those NaN cases.