Fix issue (bug) in EqualizerCore
Issue link:
https://github.com/RkShaRkz/ExoticaTechnologies/issues/50

The problem stemmed from the fact that both the "damage increase percent per 100 range decrease" was divided by 100,
but also the actual "max range adjusted ratio" representing how many hundreds of range we're over the upper range limit
was also being divided by 100.

Whether PMD's issue was that he was planning on using them as multipliers (in which case it might have been fine),
or that he just had a brainfart and scaled both values down by 100 (causing a cumulative 10000 division in the end)
is of no interest. The problem has been tracked down and rectified.

We will still keep on using the modifyPercent() due to how it stacks with other modifier-adjusting multipliers.

Here are some results from my test run, using a pure equalziercore on a ship with no other modifiers:

**BEFORE THE FIX:**
[EqualizerCore] modifying damage of weapon: aria_gustav [modifier.modifyPercent]	buffId: ET_EqualizerCore, buff: 2.5
[EqualizerCore] --> getWeaponBaseRangeFlatMod()	weaponId aria_gustav of shipId: 63f1ec6a-0bb2-4fd2-b18c-be1eab5f46b5 has base maxRange: 1800.0
[EqualizerCore] getWeaponBaseRangeFlatMod()	getUpperRangeLimit(): 800.0	getLowerRangeLimit(): 550.0
[EqualizerCore] <-- getWeaponBaseRangeFlatMod()	weaponId aria_gustav of shipId: 63f1ec6a-0bb2-4fd2-b18c-be1eab5f46b5 returning modified baseRangeMod -500.0		for adjusted maxRange: 1300.0

**AFTER THE FIX:**
[EqualizerCore] --> getWeaponBaseRangeFlatMod()	weaponId aria_gustav of shipId: 4ce05655-b0c9-4c85-9f5b-54130cc0380b has base maxRange: 1800.0
[EqualizerCore] getWeaponBaseRangeFlatMod()	getUpperRangeLimit(): 800.0	getLowerRangeLimit(): 550.0
[EqualizerCore] <-- getWeaponBaseRangeFlatMod()	weaponId aria_gustav of shipId: 4ce05655-b0c9-4c85-9f5b-54130cc0380b returning modified baseRangeMod -500.0		for adjusted maxRange: 1300.0

[EqualizerCore] modifying damage of weapon: aria_gustav	rangeDecreaseDamageIncrease: 25.0, maxRange: 10.0, buff: 250.0
[EqualizerCore] modifying damage of weapon: aria_gustav [modifier.modifyPercent]	buffId: ET_EqualizerCore, buff: 250.0	damage: 1890.0 [BEFORE]
[EqualizerCore] modifying damage of weapon: aria_gustav	modified damage: 5640.0 [AFTER]