Slight rework (buff) to PlasmaFluxCatalyst - make multipliers work on limits too
Before the rework, the capacitor/vents limit would always remain the same: `hullsize max / 3`

The positive/negative mults weren't exactly being applied to the limits at all, but only effects,
meaning for a capital (which has e.g. 75 max vents/caps) the limit would always be 25 and then different
kinds of exoticas with their different positive/negative mults would only affect the effect granted -
200 flux for normal or 500 for pure.

The rework consists of simply making the code a bit clearer and readible (no longer having these condensed
oneliners doing too much) but actually making it easy to skim through and follow.

The positiveMult is now applied to the limit as well, and the negativeMult is applied to the coefficient so that
normal PlasmaFluxCatalyst on a capital has a max caps/vents of 25, but a pure one has a max caps/vents of 63 !!!

The underlying math didn't change at all, only the the time when the positive/negative mult was changes to that it
now also affects the limit and not just their effect. Different types of exotics certainly have different functionality,
which is what this buff/rework was all about.

NOTE TO FUTURE SELF:
when I say "positiveMult is now applied to the limit as well" I mean "instead of applying it only at effect-site, I now apply
it at limit-site and no longer apply it at effect-site". Only the timing of applying the positive/negative mult was changed,
and the code is validated for correctness so that the positive/negative mults aren't applied more than once.

So the "limit" is no longer stuck at 25 for capitals with varying levels of effects, but rather it's applied at limit and the
effects remain the same; so that the final calculation is unchanged. Except for the CR penalty of course, due to the limits being
changed.

Besides that, the code has been refactored for readability, with zero functionality changes - except that the positive/negative mult
now affect the limit, rather than just the effect. Should play nice with the additional ExoticTypes I had in mind.