Further refactoring, extracting and greatly simplifying the upgrade purchasing process
Well, alot of code was moved and extracted to BandwidthHandler. Notable new things are:
`BandwidthHandler::isAbleToPayForNextBandwidthUpgrade()` - calculates the price and checks whether
the member's fleet can cover it.
`BandwidthHandler::performNextBandwidthUpgrade()` - calculates price, checks whether the member's fleet
can afford it, performs the upgrade if it can be afforded (or is a forced upgrade) and returns whether
the upgrade was performed successfully or not

The ShipHeaderUIPlugin::doMaxBandwidthUpgrade() is greatly simplified now. It is literally
looking like "while can aford next upgrade - perform next upgrade" and all overcharging nonsense
is now removed.