Simplify ShipHeaderUIPlugin::doBandwidthUpgrade()
Now that we have all these extracted utility methods, there's zero reason
to keep doing things manually. Offload the whole work to them and BandwidthHandler.

Make the `BandwidthHandler::performNextBandwidthUpgrade()` implicitly check whether we can
afford the next upgrade, and deduce cash then upgrade if we can; or do nothing if we can't
afford it.

Much better than having this method do everything.

Time is closing for the rulescmd that shows a fleet picker on which we'll try to max out bandwidth.