Fix the bandwidth desync issue - showing of stale/no longer valid bandwidth data
When we purchase bandwidth upgrades one-by-one, the method for updating the label is called
inside the purchasing.

However, when we press the max out, the label on the panel below the ShipHeaderUIPlugin can
remain stale, causing a desync between the value shown there and the value shown above in the header.

Fix the OverviewPanelUIPlugin to track the bandwidth so that there is no desync/stale situation
happening anymore.