Making of "max bandwidth upgrade price" prognosis - WIP
The current implementation, even the first one was quite alright.
However, on the ship I'm testing with, there is a small discrepency -
STARTING CASH: 50,000,000
PROJECTED PRICE: 14,221,023
ACTUAL CHARGED PRICE: 14,221,032

This charged price is the same regardless of whether using the "purchase MAX" or
"purchase" button until max bandwidth is reached.

There is a consistent discrepency of 9 credits, which I suspect are either due to floats
(which tests confirmed as untrue) or the game doing rounding *somewhere*

So I'm investigating. Next idea is to look at the credit after each level and see
where the error shows up, and how we end up losing these 9 credits.