Issue34 prognosis (#61)
This PR contains the whole "prognosis" and the subsequent messing up of it, only to eventually realize that it's either the `MutableValue` or the `float` messing it up.

Depending on the amount of cash i'm testing with, the drift will be higher, but when testing with 15 mil - the drift is zero.

So, calling it done. Just need to clean up junk before merging it in.
There was one commit during which the drift was consistently 9 while testing with 50mil, but then I think i bumped it up to 16 by just putting synchronization blocks (and barriers) however, seeing how it works flawlessly with smaller cash sums - i'm not gonna go deeper into this. There is no overcharging on my side.

"max bandwidth upgrade" cost prognosis is effectivelly done. just needs cleanup.
---------


* Making of "max bandwidth upgrade price" prognosis - WIP

The current implementation, even the first one was quite alright.
However, on the ship I'm testing with, there is a small discrepency -
STARTING CASH: 50,000,000
PROJECTED PRICE: 14,221,023
ACTUAL CHARGED PRICE: 14,221,032

This charged price is the same regardless of whether using the "purchase MAX" or
"purchase" button until max bandwidth is reached.

There is a consistent discrepency of 9 credits, which I suspect are either due to floats
(which tests confirmed as untrue) or the game doing rounding *somewhere*

So I'm investigating. Next idea is to look at the credit after each level and see
where the error shows up, and how we end up losing these 9 credits.

* I managed to increase the drift to 16 somehow...

Well, this isn't good, even though it's more safe. so obviously i'm doing something wrong, and should stick with just 9.
wasted a whole day on this shit.

* Cleanup - part1

* Cleanup - part2