Slight tweak to HullmodExotic - update variant if on refit screen
After the onInstall and onRemove callbacks, inside the methods
`installThisHullmodExoticToFleetMembersVariant` and `unapplyExoticHullmodAndRemoveExoticaAndHullmod`
we will also call the new StarsectorAPIInteractor::updateVariantIfOnRefitScreen() that will
check whether we are on refit and tell the RefitButtonAdder to update the variant.

Perhaps this lowers the chances/occurances of the alpha subcore-related problems?
I suppose it won't hurt.