rules.csv triggering the OnDockedListener - part1
This commit actually has all it needs, this works good. I just figured to separate it into pieces.

This commit adds a small rules.csv modification to trigger the OnDockedListener when the player docks;
from there we will run through all of the FleetMembers in the player's fleet, promote their FleetData to
ETFleetData - which will have a onPreScuttle() and onPostScuttle() callbacks - and go back to life as usual.

Then, when the ship is scuttled, we can "intercept" or react to that and give back all of it's chips to the player.

**THEN** this whole thing needs to be behind a lunasetting switch to enable/disable it.