Promoting regular FleetData to ETFleetData - part2
As we go through each FleetMember of the player's fleet, we promote each of their
FleetData instance with our new ETFleetData which fires the callbacks before and after
calling super.scuttle(), thus effectivelly hooking before/after the method call...

However, the original FleetData was full of stuff, and this new is rather empty.
We will need to populate it from the original's clone.