Rename to ShipRepulsorSystem and fix push out effect
The enemy ships are now pushed directly back. I would like to add a debilitating/paralyzing effect
as well, after the ship is pushed out so that it is also spun around until it stabilizes for a while.

NEED TO RE-TEST GUARDIAN SHIELD AND MAKE SURE NOTHING BROKE

Other than that, also need to retest the case when i'm trying to push out a station, so that it should
push me away rather than maybe attract me.

tl;dr the ShipRepulsorSystem is kinda done?

- test guardian shield
- test against station and make sure you get pushed out and not the station
- add debilitating spinning effect to non-station enemy ships.
- maybe see about tweaking the visual flair a bit; while this is good the stage1 at 50 range isn't super visible.
see about making these ranges somewhat dynamic and dependant on either the <actual radius> or collision radius of the ship.
or both.