Updated strings, this seems decent
However, the "hyper-spinning" issue that is sometimes applied to myself in
simulation still remains, and it has nothing to do with the ForceApplier's
implicit momentum upscaling by a factor of 100.

In fact, when that bit is parametrized, the ships aren't pushed back much
(if at all) meaning it should remain as-is and the issue should be hunted
somewhere else.

It does seem to happen only when I push out alot of ships though...

Should probably test more in actual combat than simulation.