Seems I got rid of the "spinning" problem
I still don't quite know how it got applied to my own ship - BUT - the
ForceApplier had a part where it was modifying the angular velocity.

Once that was parametrized and we used 'false' with the ForceApplier "push out"
part of the code - I either didn't repro it, or I fixed it.

Also, I lowered the scaled rotational momentum from [-2160,2160] to [-1080,1080].

Needs cleanup, needs getting rid of superfluous logs.