Started working on the rest of the stuff - and I verified that I broke it again
Well, the scale factor aside, it seems that the `modifyAngularVelocity` does in fact prevent the problem.

Or - I just didn't manage to repro it a few times in a row. Lets investigate a bit more.

The "target" thing is underway. the whole "enemies only" / "enemies and allies" / "allies only" with the new
useless util function :)