Re-fix the spinning problem
The `modifyAngularVelocity = false` really is key in the `ForceApplier.applyMomentum(...)` call.

It prevents the spinning issue from happening.

Also, some more work was done on the enemies/enemies+allies/allies front, and that feature might as well
be done; needs testing.

However, the smaller scale was nicer, so it should probably be brought back ...