Fix kotlin setup and get rid of some unnecessary dependencies (from project)
After accidentally nuking the .idea folder without backing it up, i was stuck in project
setup hell for some 2h where kotlin could not build, had 28k errors on everything, was using 1.9
instead of 1.6 or some other bullshit...

Reverting the files did nothing. Things had to be changed from scratch and be put in again.

So here we are. project is building again. Time to test the new Swirl ...