Add remapping of user-centric angles into trigonometry angles
Something about this is super-familiar with the darts game, but the jist of the commit is this:

As a caller of the generateSwirl() method, I want to pass the ringRotations as eg [0, 30, 60, 90, 120, 150, 180],
expecting each ring after the first to rotate 30-degrees in a clockwise fashion.

However, trigonometry doesn't quite dig that, and it interprets them as counterclockwise rotations instead.

To avoid having to fix this at call-site, and to make the caller aware that he's dealing with trigonometry instead of
plain angles in degrees, it's best to just remap all user-centric angles into trigonometry-centric angles during generation
instead.

EDIT: update javadoc to tell that ring rotations should be considered CLOCKWISE (CW) instead of trigonometry-default counter clockwise (CCW)