Segments are still not being drawn .....
Even though I made sure that everything is done in degrees before being converted to radians
- which shouldn't have been the root cause since the segments might have been drawing fine
while everything was being converted to radians individually and then summed - the segments are still
not being drawn and are just dots.

Something got messed up here... i need to take a step back, and look into the code that was drawing the segments
correctly, and see whats there to see.