Adding Vector2f tests to validate the new vector math i'm making and using ...
The particles or the particle segments seem to only be on the right side of the ship.
Besides having to make these rotations CW/CCW and not only CCW for positive angles, I need
to figure out why all of the particles aren't there.

The tests seem to be passing, so the problem might not be in the Vector math part, but the
Swirl generating/drawing part ...

EDIT: added more tests for odd cases. they're all passing too ...