LOGARITHMIC one looks much better but ... (problems)
1) it's only on the 'right side' of the ship, or rather in Q1 and Q4
This might be due to the FastTrigUtils covering only -pi,pi range, which is -90 to 90.

This would explain why the segments start from Q4 (-90) and go up to Q1 (90).

Instead i should try fixing it to cover the whole -2pi, 2pi range.... if it makes sense.

2) the arcs are (vertically) offset from the center of the ship.

The arcin and swirling is all nice, however the arcs are centered around a point above my ship.
I suspect reusing the 'center' for logarithmic might be the problem, and maybe i should just use null
like i did for BEZIER.