Fix the vertical pivot for LOGARITHMIC swirl type
The problem was that we started off from translating the coord system from (0,0)
to (Cx, Cy) but never translated back - that was causing the vertical pivot.

Once the retranslation was added, the points draw nicely along the side of the ship.

HOWEVER - the problem of the points still "originating from the right side of the ship",
or rather from -pi/4 to pi/4 and remain just in Q1 and Q4 is still present and needs to be
tackled separately ...