Turns out LOGARITHMIC generation and math works out just fine
However, when trying to draw them without bumping the limits only a small portion
of the "swirl arms" were being drawn - 18 to be exact with 16 particles per segment.
When the particles were lowered to 8, more swirl arms were being drawn. When that was
lowered down to 4, they were looking like crap but even more swirl arms were drawn...

So then I realized there's gotta be some builtin limit to how many particles can be drawn.
After bumping the limit temporarily before drawing all of them, the whole swirl was drawing
quite nicely.

So, thats when i got the idea to "draw out" the swirl arms to be drawn over time.
The swirl generation should return a pair of points - the ring points, as well as the particle points.

So that Swirl::draw() can draw the rings and the emp arcs, and Swirl::drawParticles(amount) can draw the
particles over time.

That's the next step to be implemented.