Drawing over time is done
Right now, with interval of 0.1 it looks kinda bleh, but once the interval is shortened
and a few other knobs have been adjusted (e.g. drawing 2 particles per interval) and a new
MultiIntervalUtil is done (so that we can do more than one interval with a single `advance()`)
it will probably look alot better.

Still, it's worth saving in this state.