Changed to drain 6 particles per interval and tiny improvements
So, the 6 particle drain looks okay i suppose. Or at least when tested with the ork_superdread
which is huge. I need to test with smaller ships.

Some minor refactor happened on the ShipAttractorSystem side, but code is ready for the new
"continuousDrain" parameter which will pre-drain arms so that we get a continuously-shrinking
swirl on first circle.