Implement 'continuousDrain' option
While it doesn't quite hit close to what I originally had in mind, now that I think about it,
it does greatly reduce the vortex duration, it still lasts for 2-3 seconds, and looks decent on
both huge capitals and small destroyers. So I'll keep it.

The original idea was something like this:
After the first arm of eg 90 points draws one point and removes it, the second arm should
pre-remove 1 point before drawing, effectivelly removing 2 points for that iteration,
the third arm should pre-remove 2 points before drawing (effectivelly removing 3 points),
the fourth arm should pre-remove 3 points ...

This is all under assumption that each arm will draw only one point. In case we're drawing multiple,
we should increment each time so that the arm to-be-drawn-next starts off from the size we finished at.

This causes the "continuous drain" and visible shrinking in the first circle of the swirl.

The thing I implemented now *doesn't quite work like that* but is somewhere close. And doesn't look bad.
So I'll just leave a TODO for my future self, never look at it and be fine with it :D

EDIT: actually left a FIXME
EDIT2: actually brought back the old TODO so i have a clue what I meant by this during code-review ...