Abandon the safe fireAll() method in favor of FireAllKotlin
The FireAllKotlin is the new object that inherits FireAll, and it's static
fire method calls into the new fire_safe extension method, which then calls
into the old FireAll.fire() static method.

This is done this way because I cannot make static extension methods on java objects.
*BUT* - I can do it this way, and since every FireAll.fire() invocation essentially creates
a new FireAll instance to call it on, having one extra instance with my object should be fine.

I think.

This is essentially the old `fireAll()` method generalized a bit more so it can be reused from
other places too.