Initial commit of ShipAttractorSystem - works kinda good so far
The initial implementation of the ShipAttractorSystem which depends on the hull size and number of modules
has been implemented, which also starts from a constant PULL_IN_STRENGTH of 1000.

Small ship can kinda slightly pull in larger ships, large ships can really pull in other ships, so for now I doubt
there needs to be a target-specific mass thing plugged into the equation.