New AngleDegreeType enum and a new parameter added to `generateSwirl()`
Before this, the `ringRotationsDegrees` parameter was *always* expected to be
in user-centric (azimuth/bearing) system. With the new `ringRotationDegreesListAngleType`
parameter, it can be set to be either AngleDegreeType.USER_CENTRIC or AngleDegreeType.TRIGONOMETRIC,
and we will do no remapping for TRIGONOMETRIC degree angles.

This enum will be more useful for the next system, when we get to arc generation code.